added a remove image button
Now if there are images you decide won't work well, then they can be easily removed from the data file
added a note that reads out in the console when a save occurs to give me better peace of mind.
it might be nice to add and ability to easily undo an image removal or to add data to the set.